feat: 장소를 관광지와 음식점으로 나누어 보여줌
